python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.118.0
flake8==7.3.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
isort==6.1.0
jiter==0.11.0
mccabe==0.7.0
mypy_extensions==1.1.0
openai==1.109.1
openai-agents==0.3.3
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pycodestyle==2.14.0
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
pytokens==0.1.10
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
zipp==3.23.0
slowapi==0.1.9
tenacity==9.0.0
cachetools==5.5.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Pin to compatible version with passlib 1.7.4

# Monitoring & Logging
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0

# Templating
Jinja2==3.1.4

# Testing dependencies
pytest==8.4.2
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-xdist==3.6.0
faker==33.1.0
